requirements.txt
crewai
crewai_tools
langchain_community==0.0.29
python-dotenv==1.0.0   
